hal: check if separate backend is used for voice headset
If conc. capture is disabled, we select voice headphones irrespective
of if whether separate backend is used for voice headset or not

Add check all backends match API to support both capture and playback
devices, and use it to check if separate backend is added for
headset-mic and select the device accordingly.

Change-Id: I405b52fbf14ca800bd188a62ba83bf82aa61ea1f
Signed-off-by: dotkit <ewprjkt@proton.me>